requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pymongo
passlib[bcrypt]
pyjwt
python-dotenv
pydantic[email]
PyMuPDF
google-generativeai
pdfplumber
Pillow
imagehash
docling
docling-core
chromadb
sentence-transformers
crewai
langchain-google-genai
ollama
